requirements.txt
opencv-python>=4.7.0.72
Flask>=2.2.3
numpy>=1.24.2
scipy>=1.10.1
Pillow>=9.5.0